Chore: simplified nuxt-module-ipfs module loading
