+++ modules/nuxt-module-ipfs/plugin.js
